+++ libtfhe/include/Encryptor.hpp
@@ -19,7 +19,7 @@ namespace TFHE {
         CipherText<SZ>
         encrypt(uint64_t msg) const {
             CipherText<SZ> text(_param);
-            for (int i = 0; i < SZ; ++i)
+            for (size_t i = 0; i < SZ; ++i)
                 bootsSymEncrypt(text._sample.get() + i, msg & 1ull, _key._key.get()), msg >>= 1;
             return text;
         }
@@ -27,7 +27,7 @@ namespace TFHE {
         uint64_t
         decrypt(const CipherText<SZ> &cip) const {
             uint64_t res = 0;
-            for (int i = SZ - 1; i >= 0; --i)
+            for (size_t i = SZ - 1; i >= 0; --i)
                 res <<= 1, res |= bootsSymDecrypt(cip._sample.get() + i, _key._key.get());
             return res;
         }
